Refactor fixture location. Only want one place to look
